% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Installer.R
\name{column.rm}
\alias{column.rm}
\title{This function removes columns from a table
pass in the df and the column and it will return a column without the mentioned column(s)}
\usage{
column.rm(df, col_name)
}
\arguments{
\item{df}{a data frame}

\item{col_name}{1 or multiple concatenated column names}
}
\value{
returns the data frame without the specified columns
}
\description{
This function removes columns from a table
pass in the df and the column and it will return a column without the mentioned column(s)
}

