% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_pkgs.R
\name{dev.pkgs}
\alias{dev.pkgs}
\title{This function is a second package installer, used only by developers}
\usage{
dev.pkgs()
}
\description{
This function is a second package installer, used only by developers
}

