% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Installer.R
\name{df.c2f}
\alias{df.c2f}
\title{This function converts the character columns into factor, useful in visualising correlation matrix}
\usage{
df.c2f(df)
}
\description{
This function converts the character columns into factor, useful in visualising correlation matrix
}
