% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Installer.R
\name{df.class.extract}
\alias{df.class.extract}
\title{This function extracts only columns that are either numeric, character or factor}
\usage{
df.class.extract(df, class_type)
}
\arguments{
\item{class_type}{either "numeric", "character" or "factor"}
}
\description{
This function extracts only columns that are either numeric, character or factor
}

