% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Installer.R
\name{df.col.select}
\alias{df.col.select}
\title{This function extracts only columns references that are either numeric, character or factor}
\usage{
df.col.select(df, class_type)
}
\arguments{
\item{class_type}{either "numeric", "character" or "factor"}
}
\description{
This function extracts only columns references that are either numeric, character or factor
}
