% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Installer.R
\name{dt.double_quote_fix}
\alias{dt.double_quote_fix}
\title{When read in CSVs using fread, " usually being read in as "", this function fixes it}
\usage{
dt.double_quote_fix(temp_col)
}
\arguments{
\item{temp_col}{column to perform this operation}
}
\description{
When read in CSVs using fread, " usually being read in as "", this function fixes it
}
