% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Installer.R
\name{f2c}
\alias{f2c}
\title{This function converts factor column to char
#' @param col the column in a data frame}
\usage{
f2c(col)
}
\description{
This function converts factor column to char
#' @param col the column in a data frame
}
\examples{
x$abc <- f2c(x$abc)
}

