% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Installer.R
\name{f2n}
\alias{f2n}
\title{This function converts factor column to numeric (usually applicable to year)}
\usage{
f2n(col)
}
\arguments{
\item{col}{the column in a data frame}
}
\description{
This function converts factor column to numeric (usually applicable to year)
}
\examples{
x$abc <- f2n(x$abc)
}
