% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Installer.R
\name{fn.is.dt.end}
\alias{fn.is.dt.end}
\title{complimentary to the function above}
\usage{
fn.is.dt.end(df, dt_flag)
}
\arguments{
\item{dt_flag}{is a variable produced in the function above}
}
\description{
complimentary to the function above
}

