% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Installer.R
\name{fn.is.dt.start}
\alias{fn.is.dt.start}
\title{This function is mainly for internal use so I wont export it right now
Checks if the table is a df or a dt}
\usage{
fn.is.dt.start(df)
}
\arguments{
\item{df}{a data.frame or data.table}
}
\description{
This function is mainly for internal use so I wont export it right now
Checks if the table is a df or a dt
}

