% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Installer.R
\name{fn.piper}
\alias{fn.piper}
\title{This is a piper for functions}
\usage{
fn.piper(...)
}
\arguments{
\item{...}{the functions you want to pipe}
}
\value{
returns a custom function that executes all the functions in reversed order
}
\description{
This is a piper for functions
}
\examples{
#' custom_fn <- fn.piper(length, unique)
custom_fn(rep(1:5, 1:5)) would be 5

custom_fn <- fn.piper(is.numeric, length, unique)
custom_fn(rep(1:5, 1:5)) would be TRUE

}
