% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case_insensitive_joins.R
\name{insensitive.join}
\alias{insensitive.join}
\title{This performs case insensitive joins of the joins in dplyr
https://gist.github.com/jimhester/a060323a05b40c6ada34}
\usage{
insensitive.join(fun = inner_join)
}
\description{
This performs case insensitive joins of the joins in dplyr
https://gist.github.com/jimhester/a060323a05b40c6ada34
}
\examples{
df <- insensitive.join(left_join)(df_x, df_y, by = list(x = c(), y = c()))
}
