% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Installer.R
\name{range_gen}
\alias{range_gen}
\title{This function generates a range based on a column and a specified range, remember to change the new column name to your choice}
\usage{
range_gen(df, col_ref, range)
}
\arguments{
\item{df}{data frame}

\item{col_ref}{double quoted column reference}

\item{range}{the range that you want to generate}
}
\description{
This function generates a range based on a column and a specified range, remember to change the new column name to your choice
}
\examples{
col_ref = "Size", range = 5. this will find the column Size in the df supplied and generate size ranges of 5 inch, 25-30, 30-35
}
