% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbind_fill_DT.R
\name{rbind.fill.DT}
\alias{rbind.fill.DT}
\title{This function does an rbind.fill / smartbind on a list of data.frames}
\usage{
\method{rbind}{fill.DT}(ll)
}
\arguments{
\item{ll}{list of data frames}
}
\description{
This function does an rbind.fill / smartbind on a list of data.frames
}
\examples{
huge_df <- rbind.fill.DT(list_of_dfs)
}
