% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_large_amount_of_sheets.R
\name{read.dem.sheets}
\alias{read.dem.sheets}
\title{this function is short line to read multiple sheets in a df}
\usage{
read.dem.sheets(file_path, sheetnames)
}
\arguments{
\item{file_path}{path of .xlsx files}

\item{sheetnames}{an array of sheet names you want to read}
}
\value{
returns a list of data.frames of the sheets read
}
\description{
this function is short line to read multiple sheets in a df
}
