% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Installer.R
\name{save.xlsx}
\alias{save.xlsx}
\title{This function saves multiple dfs into one workbook
The sheet names will be the variable names}
\usage{
save.xlsx(file, ...)
}
\arguments{
\item{file}{file path you wish to write to}

\item{...}{list of dfs you wish to write}
}
\description{
This function saves multiple dfs into one workbook
The sheet names will be the variable names
}
