% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serial_joiner.R
\name{serial.joiner}
\alias{serial.joiner}
\title{this function join multiple dfs using dplyr}
\usage{
serial.joiner(fun = inner_join, list_of_df)
}
\arguments{
\item{fun}{types of join you wish to perform}

\item{list_of_df}{a number of dfs you require to join}
}
\description{
this function join multiple dfs using dplyr
}
\examples{
df <- insensitive.join(left_join, list(df_x,df_y,df_z))
}
