% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Installer.R
\name{size.encoder}
\alias{size.encoder}
\title{Sometimes when reading in < and >, it gets encoded differently, this function fixes it}
\usage{
size.encoder(temp_col)
}
\arguments{
\item{temp_col}{column to perform this operation}
}
\description{
Sometimes when reading in < and >, it gets encoded differently, this function fixes it
}
