% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Installer.R
\name{size_cat_gen}
\alias{size_cat_gen}
\title{This function generates size range}
\usage{
size_cat_gen(df, col_name, range)
}
\arguments{
\item{df}{df you want to generate size cat}

\item{col_name}{Size column name of in df you want to generate range for}

\item{range}{the window of size range, eg. 5 means say 10"-15" & 15"-20"}
}
\description{
This function generates size range
}

