% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Installer.R
\name{source_https}
\alias{source_https}
\title{This function sources scripts from git_hub}
\usage{
source_https(url, ...)
}
\arguments{
\item{url}{urls}
}
\description{
This function sources scripts from git_hub
}
\examples{
source_https("https://raw.github.com/tonybreyal/Blog-Reference-Functions/master/R/bingSearchXScraper/bingSearchXScraper.R",
"https://raw.github.com/tonybreyal/Blog-Reference-Functions/master/R/htmlToText/htmlToText.R")
}

